adding create functionality and index for menu in admin and also some error for form

+++ database/migrations/2021_01_05_183455_create_menus_table.php
@@ -19,7 +19,7 @@ class CreateMenusTable extends Migration
             $table->text('desc');
             $table->text('menu_image')->nullable();
             $table->bigInteger('quantity')->nullable();;
-            $table->bigInteger('price')->nullable();;
+            $table->float('price', 8, 2)->nullable();;
             $table->timestamps();
         });
     }
